% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertiesHandler.R
\name{check_if_properties_exist}
\alias{check_if_properties_exist}
\title{Check for the existence of the properties file.}
\usage{
check_if_properties_exist()
}
\value{
\code{TRUE} if the \code{properties.rds} exists at the root of the \code{tmcr}
directory, otherwise \code{FALSE}.
}
\description{
Check for the existence of the \code{properties.rds} file at the root of the
\code{tmcr} directory.
}
\seealso{
\code{\link{get_tmcr_directory}}, \code{\link[base]{files}}
}
