% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerciseDataHandler.R
\name{create_exercise_metadata}
\alias{create_exercise_metadata}
\title{Create a JSON file containing exercise metadata}
\usage{
create_exercise_metadata(exercise_id, exercise_directory, exercise_name)
}
\arguments{
\item{exercise_id}{ID of the exercise.}

\item{exercise_directory}{Path to the directory where the exercise is downloaded into.}

\item{exercise_name}{Name of the exercise.}
}
\value{
Either \code{NULL} or an integer status.
}
\description{
Create a \code{JSON} file containing exercise metadata.
}
\details{
Creates a \code{JSON} file in the exercise's directory root. The \code{JSON}
file contains the exercise's name and id, which is used when the user wants to upload
their exercise submission to the TMC server.
}
\seealso{
\code{\link[base]{file.path}}, \code{\link[base]{connections}},
\code{\link[jsonlite]{toJSON}}, \code{\link[jsonlite]{unbox}}, \code{\link[base]{cat}}
}
