% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertiesHandler.R
\name{create_properties_file}
\alias{create_properties_file}
\title{Create the properties file}
\usage{
create_properties_file(tmcr_projects = "tmcr-projects")
}
\arguments{
\item{tmcr_projects}{Name of the directory where the exercises are downloaded into.
\code{tmcr-projects} by default.}
}
\value{
Always \code{NULL}.
}
\description{
Create the \code{properties.rds} file, which includes the path to the directory
where the exercises are downloaded into.
}
\seealso{
\code{\link{get_tmcr_directory}}, \code{\link[base]{saveRDS}}
}
