% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunResultHtml.R
\name{createRunSourcingFailHtml}
\alias{createRunSourcingFailHtml}
\title{Create HTML for displaying a run or sourcing fail}
\usage{
createRunSourcingFail(runResults)
}
\arguments{
\item{runResults}{Results from the failed attempt of running tests.}
}
\value{
A single shiny HTML tag object containing all of the elements in the created HTML
run result output.
}
\description{
Create HTML for displaying a run or sourcing fail.
}
\details{
Creates an HTML view for displaying information related to a run- or
sourcing fail
}
\seealso{
\code{\link[shiny]{tags}}
}
