% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunResultHtml.R
\name{createTestResultsHtml}
\alias{createTestResultsHtml}
\title{Create HTML for displaying test results}
\usage{
createTestResultsHtml(testResults, showAll)
}
\arguments{
\item{testResults}{List of test results.}

\item{showAll}{Boolean for showing all test results.}
}
\value{
A single shiny HTML tag object containing all of the elements in the created HTML
test result output.
}
\description{
Create HTML for displaying test results
}
\details{
The created HTML includes a percentage bar for showing how many of the tests
passed, and a list of paragraph elements for each of the test results. If the user has
ticked off the \code{Show all results} checkbox, then instead of a list only a single
paragraph element is shown. The single parahraph is either of the first failing test,
or a message showing "All tests passed" if none of the tests failed. If the current
exercise doesn't have any tests associated with it, then only a single paragraph reading
"No tests for exercise" is returned.
}
\seealso{
\code{\link[shiny]{tags}}
}
