% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{deleteCredentials}
\alias{deleteCredentials}
\title{Delete the saved credentials}
\usage{
deleteCredentials()
}
\value{
True if succeeded in deleting the credentials. False if the deletion did not succeed or the saved
credentials were not found.
}
\description{
Delete the saved credentials.
}
