% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{disable_elements}
\alias{disable_elements}
\title{Disable shiny input elements}
\usage{
disable_elements(...)
}
\arguments{
\item{...}{IDs of the shiny input elements which are to be disabled.}
}
\description{
Disable shiny input elements.
}
\details{
Disables the shiny input elements and sets the global boolean \code{UI_disabled}
to \code{FALSE}. This is done in order to prevent excessive buffering of button actions if
the user starts clicking buttons in other tabs during an operation that is taking a long itme.
}
\examples{
disable_elements("login", "logout", "username", "password")
}
