% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{disable_submit_tab}
\alias{disable_submit_tab}
\title{Disable Shiny input elements in the Test & Submit tab}
\usage{
disable_submit_tab()
}
\description{
Disable \code{Shiny} input elements in the Test & Submit tab
}
\seealso{
\code{\link{disable_elements}}
}
