% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{disable_tab_UI_list}
\alias{disable_tab_UI_list}
\title{Disable Shiny input elements using UI tab list}
\usage{
disable_tab_UI_list(tab)
}
\arguments{
\item{tab}{A named input element vectors or NULL.}
}
\description{
Disable \code{Shiny} input elements given by UI list
}
\details{
Disables the \code{Shiny} input elements.
This is done in order to prevent button actions while an
operation that might take a long itme is running.
}
\seealso{
\code{\link{enable_UI_elements}}
}
