% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{downloadedExercises}
\alias{downloadedExercises}
\title{Get a list of downloaded exercises}
\usage{
downloadedExercises()
}
\value{
List containing the names of the exercises.
}
\description{
Get a list of the names of downloaded exercises.
}
