% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{downloadedExercisesPaths}
\alias{downloadedExercisesPaths}
\title{Get a list of downloaded exercises}
\usage{
downloadedExercisesPaths()
}
\value{
List of downloaded exercises as named vector \code{c(name = path)}format.
}
\description{
Get a list of downloaded exercises as named vector \code{c(name = path)}format.
}
\details{
Searches for downloaded exercise projects at the TMC R project folder.
}
\seealso{
\code{\link{findExercisesFromPath}}, \code{\link{get_projects_folder}},
\code{\link{getExerciseName}}, \code{\link[stats]{setNames}}
}
