% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_course_tab}
\alias{enable_course_tab}
\title{Enable Shiny input elements in the Course & Exercise tab}
\usage{
enable_course_tab()
}
\description{
Enable \code{Shiny} input elements in the Course & Exercise tab.
}
\seealso{
\code{\link{enable_elements}}
}
