% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_elements}
\alias{enable_elements}
\title{Enable shiny input elements}
\usage{
enable_elements(...)
}
\arguments{
\item{...}{IDs of the shiny input elements which are to be enabled.}
}
\description{
Enable shiny input elements.
}
\details{
Enables the shiny input elements and sets the global boolean \code{UI_disabled} to
\code{FALSE} after 1000 milliseconds (=1 second) have passed. This is done in order to
prevent excessive buffering of button actions if the user starts clicking buttons in other
tabs during an operation that is taking a long itme.
}
\examples{
enable_elements("login", "logout", "username", "password")
}
