% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_login_tab}
\alias{enable_login_tab}
\title{Enable Shiny input elements in the Login tab.}
\usage{
enable_login_tab()
}
\description{
Enable \code{Shiny} input elements in the Login tab.
}
\seealso{
\code{\link{enable_elements}}
}
