% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_submit_tab}
\alias{enable_submit_tab}
\title{Enable Shiny input elements in the Test & Submit tab}
\usage{
enable_submit_tab()
}
\description{
Enable \code{Shiny} input elements in the Test & Submit tab
}
\seealso{
\code{\link{enable_elements}}
}
