% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{exerciseFromWd}
\alias{exerciseFromWd}
\title{Get the currently open exercise}
\usage{
exerciseFromWd()
}
\value{
Name of the exercise directory. If the current working directory isn't an exercise folder, returns an
empty string.
}
\description{
Get the currently open exercise (the exercise which is located at the current working directory).
}
\details{
Checks if the current working directory is a subfolder in the TMC projects -folder.
}
