% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{exercisePathFromWd}
\alias{exercisePathFromWd}
\title{Get the currently open exercise}
\usage{
exercisePathFromWd()
}
\value{
Exercise as a named vector: \code{c(name = path)}. If the
current working directory isn't a valid exercise project folder,
returns \code{c(" ", " ")}.
}
\description{
If the currently open exercise (the exercise which is
located at the current working directory) is a valid TMC R exercise
project, return it as a \code{c(name = path)} vector.
}
\details{
The exercise is a valid TMC R exercise project if it
contains the \code{R} and \code{tests/testthat} folders and is
located in the TMC R projects directory.
}
\seealso{
\code{\link[base]{basename}}, \code{\link[base]{getwd}},
\code{\link{get_projects_folder}} \code{\link{pathIsExercise}},
\code{\link{getExerciseName}}, \code{\link[stats]{setNames}}
}
