% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{fetchClientIdAndSecret}
\alias{fetchClientIdAndSecret}
\title{Fetch client id and secret from the server}
\usage{
fetchClientIdAndSecret(serverAddress)
}
\arguments{
\item{serverAddress}{Address of the TMC server which the user wants to log in to.}
}
\value{
Client ID and secret fetched from the server.
}
\description{
Fetch the client id and secret required in OAuth2 authentication from the server.
}
\details{
Called from the \code{\link{authenticate}} function which provides the desired TMC server address.
}
\seealso{
\code{\link[=func]{authenticate}}
}
