% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{findExercisesFromPath}
\alias{findExercisesFromPath}
\title{Get paths to exercises}
\usage{
findExercisesFromPath(path)
}
\arguments{
\item{path}{Path to the directory where the exercises are searched from.}
}
\value{
A vector of full file paths to the found exercises. Always contains an
empty \code{""} path.
}
\description{
Recursively searches for exercise projects in the given file path.
}
\seealso{
\code{\link[base]{list.dirs}}, \code{\link{pathIsExercise}}
}
