% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{get_all_courses}
\alias{get_all_courses}
\title{Get all courses offered by a TMC organization}
\usage{
get_all_courses(organization, credentials)
}
\arguments{
\item{organization}{Organization slug (Identifying URL-friendly name).}

\item{credentials}{List of user credentials.}
}
\value{
List of course names and course ids belonging to the given
organization.  If \code{credentials} are not proper or sending the
\code{HTTP-GET} request failed, returns a list with 2 empty sublists
called \code{id} and \code{name}.
}
\description{
Get all courses offered by a TMC organization.
}
\details{
Reads the \code{OAuth2} token and server address from
\code{.credentials.json} and uses them to make a \code{HTTP-GET}
request for the list of courses belonging to the organization,
given the user has proper credentials.
}
\seealso{
\code{\link{getCredentials}},
\code{\link[httr]{stop_for_status}}, \code{\link[jsonlite]{fromJSON}}
}
