% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{get_all_exercises}
\alias{get_all_exercises}
\title{Get all exercises of a TMC course}
\usage{
get_all_exercises(course, credentials)
}
\arguments{
\item{course}{ID of the course.}

\item{credentials}{List of user credentials.}
}
\value{
List of exercises in the course. If
\code{credentials} are not proper or sending the \code{HTTP-GET} request
failed, returns an empty list.
}
\description{
Get all exercises of a TMC course.
}
\details{
Reads the \code{OAuth2} token and server address from
\code{credentials} and uses them to make a \code{HTTP-GET}
request for the list of exercises in the course.
}
\seealso{
\code{\link{getCredentials}},
\code{\link[httr]{stop_for_status}}, \code{\link[jsonlite]{fromJSON}}
}
