% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{get_all_organizations}
\alias{get_all_organizations}
\title{Get all TMC organizations}
\usage{
get_all_organizations(credentials)
}
\arguments{
\item{credentials}{List of user credentials.}
}
\value{
List of TMC organization names and slugs. If
\code{credentials} are not proper or sending the \code{HTTP-GET} request
failed, returns a list with 2 empty sublists called \code{name}
and \code{slug}.
}
\description{
Get all TMC organizations.
}
\details{
Reads the \code{OAuth2} token and server address from
\code{credentials} and uses them to make a \code{HTTP-GET}
request for the list of organizations.
}
\seealso{
\code{\link{getCredentials}},
\code{\link[httr]{stop_for_status}}, \code{\link[httr]{add_headers}},
\code{\link[jsonlite]{fromJSON}}
}
