% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerciseDataHandler.R
\name{get_exercise_metadata}
\alias{get_exercise_metadata}
\title{Read exercise metadata}
\usage{
get_exercise_metadata(exercise_path)
}
\arguments{
\item{exercise_path}{File path to the exercise project folder.}
}
\value{
\code{R} object created from \code{.metadata.json}. \code{NULL} if
\code{.metadata.json} doesn't exist.
}
\description{
Read exercise metadata from the \code{.metadata.json} file.
}
\seealso{
\code{\link[base]{file.path}}, \code{\link[base]{files}},
\code{\link[jsonlite]{read_json}}
}
