% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{get_json_from_submission_url}
\alias{get_json_from_submission_url}
\title{Get exercise submission result JSON}
\usage{
get_json_from_submission_url(response, token)
}
\arguments{
\item{response}{HTTP response to the exercise submission.}

\item{token}{OAuth2 token associated with the current login session to the TMC server.}
}
\value{
HTTP response from the TMC server containing the submission result JSON if the
server has finished processing the exercise submission. \code{NULL} if the HTTP
GET-reaction failed.
}
\description{
Get exercise submission result JSON from the TMC server.
}
\details{
Extracts the exercise submission url from the given response and calls the
\code{\link{get_submission_json}} function which does the actual HTTP GET-request.
}
\seealso{
\code{\link{get_submission_json}}
}
