% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertiesHandler.R
\name{get_projects_folder}
\alias{get_projects_folder}
\title{Get the TMC R-project directory path}
\usage{
get_projects_folder()
}
\description{
Get the file path to the \code{tmcr-projects} folder,
which contains the TMC R exercise projects.
}
\details{
TODO: Do this after PropertiesHandler.R refactoring
}
