% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertiesHandler.R
\name{get_tmcr_directory}
\alias{get_tmcr_directory}
\title{Get the tmcr directory path}
\usage{
get_tmcr_directory()
}
\value{
File path to the location of the \code{tmcr} directory
}
\description{
Get the file path to the \code{tmcr} directory path.
}
\details{
Creates a file path to the \code{tmcr} directory. It is located where the
user's \code{HOME} environment variable points to: by default to the \code{home} directory
on Linux and \code{user/documents} on Windows. If the \code{tmcr} directory does not exist
at the file path then it is created to be at that location.
}
\seealso{
\code{\link[base]{normalizePath}}, \code{\link[base]{file.path}},
\code{\link[base]{dir.exists}}
}
