% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{getAllOrganizations}
\alias{getAllOrganizations}
\title{Get all TMC organizations}
\usage{
getAllOrganizations()
}
\value{
List of TMC organization names and slugs. If reading \code{.credentials.json}
 or sending the HTTP request failed, returns a list with 2 empty sublists called \code{name}
 and \code{slug}.
}
\description{
Get all TMC organizations.
}
\details{
Reads the OAuth2 token and server address from \code{.credentials.json} and uses
them to make a HTTP-Get request for the list of organizations.
}
