% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{getCredentials}
\alias{getCredentials}
\title{Get the TMC user credentials from a file}
\usage{
getCredentials()
}
\value{
A list with the keys \code{username}, \code{token} and \code{serverAddress} and their corresponding values.
If reading the file was unsuccessful then an empty list is returned instead.
}
\description{
Get the TMC user credentials (username, access token and the server address) from the
\code{credentials.rds} file.
}
