% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{getDialogMessage}
\alias{getDialogMessage}
\title{Get message to display in submission result pop-up dialog.}
\usage{
getDialogMessage(submitResults)
}
\arguments{
\item{submitResults}{List of data read from the submission result
\code{JSON}.}
}
\value{
Message showing if submitting the exercise failed, some tests
failed or all tests passed.
}
\description{
Creates a message to be shown on the submit result
pop-up dialog from the submission results.
}
