% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{getExerciseFromServer}
\alias{getExerciseFromServer}
\title{Get the exercise submission results from the TMC server}
\usage{
getExerciseFromServer(response, token, sleepTime)
}
\arguments{
\item{response}{\code{HTTP} response to the exercise submission upload.}

\item{token}{\code{OAuth2} token associated with the current login session.}

\item{sleepTime}{The time to sleep between queries to the tmc-server.}
}
\value{
Submission result with non \code{NULL} \code{results} if
processing the tests in the TMC server was successful. List keys:
\code{results}, \code{error}. Error is not \code{NULL} if processing
the tests ended in error.
}
\description{
Get the exercise submission results from the TMC server
}
\details{
Queries the server with \code{HTTP-GET} requests until the server
has finished processing the exercise submission.
}
\seealso{
\code{\link{get_json_from_submission_url}},
\code{\link[shiny]{withProgress}}
}
