% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{getExerciseName}
\alias{getExerciseName}
\title{Get the exercise's name}
\usage{
getExerciseName(exercisePath)
}
\arguments{
\item{exercisePath}{File path to the exercise project directory.}
}
\value{
Exercise's name read from \code{.metadata.json}. If the file doesn't
exist, returns the name of the path's basename (the final directory/file in
the file path).
}
\description{
Get the name of the exercise located at the given file path.
}
\details{
Reads the \code{.metadata.json} file for the exercise name.
}
\seealso{
\code{\link{get_exercise_metadata}}, \code{\link[base]{basename}}
}
