% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{getExercisePath}
\alias{getExercisePath}
\title{Get a file path to the specified exercise}
\usage{
getExercisePath(exercise)
}
\arguments{
\item{exercise}{Name of the exercise}
}
\value{
File path to the exercise which name was given as a parameter
}
\description{
Get a file path to the specified exercise.
}
\examples{
getExercisePath("0_0_helloworld")
}
