% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{getServerAddress}
\alias{getServerAddress}
\title{Get the saved server address}
\usage{
getServerAddress()
}
\value{
The saved server from the .server file. NULL if the file does not exists or if it is corrupted.
}
\description{
Get the saved server address.
}
