% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{login}
\alias{login}
\title{Login to a TMC server}
\usage{
login(clientID, secret, username, password, serverAddress)
}
\arguments{
\item{clientID}{Client ID used in OAuth2.}

\item{secret}{Secret used in OAuth2.}

\item{username}{Username of a TMC account.}

\item{password}{Password matching the inputed username of a TMC account.}

\item{serverAddress}{Address of the TMC server which the user wants to log in to.}
}
\value{
An OAuth2 token if the authentication was succesful, otherwise returns an error message.
}
\description{
Logs in to a TMC server using an username and a password along with a clientID and secret used
for OAuth2 authentication.
}
\details{
Called from the \code{\link{authenticate}} function, which provides the client id and secret
used in this function's parameters. If logging in was successful, saves the login credentials (username,
access-token and the server address) in the credentials file.
}
\seealso{
\code{\link{authenticate}}
}
