% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{processSubmission}
\alias{processSubmission}
\title{Read test result data from the submission result JSON}
\usage{
processSubmission(output)
}
\arguments{
\item{output}{HTTP response containing the submission result JSON.}
}
\value{
List of test results received from the submission result JSON.
}
\description{
Read test result data from the submission result JSON.
}
\details{
Creates a list of test results received from the submission result JSON.
}
