% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{showMessage}
\alias{showMessage}
\title{Show submission results in a pop-up dialog box}
\usage{
showMessage(submitResults)
}
\arguments{
\item{submitResults}{List of data read from the submission result
\code{JSON}.}
}
\description{
Show submission reuslts in a pop-up dialog box.
}
\seealso{
\code{\link{getDialogMessage}},
\code{\link[rstudioapi]{showDialog}}
}
