% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{submit_exercise}
\alias{submit_exercise}
\title{Submit current exercise and process the response}
\usage{
submit_exercise(path, credentials)
}
\arguments{
\item{path}{Path to the currently chosen directory.}

\item{credentials}{List of user credentials.}
}
\value{
List of data read from the submission result \code{JSON}.
List keys: \code{tests}, \code{exercise_name},
\code{all_tests_passed}, \code{points}, \code{error}.  \code{error}
is not \code{NULL}if submitting the exercise to the server failed
}
\description{
Submit the currently chosen exercise to the TMC server
and process the \code{JSON} received from the response.
}
\details{
Submits the currently open exercise to the TMC server,
queries the server until it has finished processing the submission,
reads the data from the \code{JSON} received in the \code{HTTP}
response and shows a message popup showing if all of the tests passed
or not.
}
\seealso{
\code{\link{submit_current}},
\code{\link{processSubmissionJson}}, \code{\link{showMessage}}
}
