% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMC_plugin.R
\name{tmcGadget}
\alias{tmcGadget}
\title{Run the TMC addin}
\usage{
tmcGadget()
}
\description{
Run the TMC addin on the \code{RStudio} viewer pane.
}
\details{
The TMC \code{RStudio} addin was made using
\code{\link[shiny]{shiny-package}}, which allows making web
applications and \code{RStudio} addins using \code{R}.
}
