% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{upload_exercise}
\alias{upload_exercise}
\title{Upload an exercise to the TMC server}
\usage{
upload_exercise(token, exercise_id, project_path, server_address,
        zip_name = "temp", remove_zip = TRUE)
}
\arguments{
\item{token}{\code{OAuth2} token associated with the current login session.}

\item{exercise_id}{ID of the submitted exercise.}

\item{project_path}{File path to the directory of the submitted exercise.}

\item{server_address}{Address of the TMC server where the exercise is submitted to.}

\item{zip_name}{Name of the \code{zip} file which contains the exercise submission. Default is
\code{temp}}

\item{remove_zip}{\code{TRUE} or \code{FALSE} depending on if you wish to delete the
submission \code{zip} file after sending it to the server. Defaults to \code{TRUE}.}
}
\value{
\code{HTTP} response to the submission attempt.
}
\description{
Upload an exercise to the TMC server
}
\details{
Packs the exercise directory into a \code{zip} file and sends it to the TMC server.
}
\examples{
upload_exercise(token = token, exercise_id = 36463, project_path = getwd(),
           server_address = "https://tmc.mooc.fi")

}
\seealso{
\code{\link[httr]{add_headers}}, \code{\link[httr]{upload_file}},
\code{\link[httr]{POST}}, \code{\link[base]{file.remove}}
}
