% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerciseDataHandler.R
\name{create_exercise_metadata}
\alias{create_exercise_metadata}
\title{Create a JSON file containing exercise metadata}
\usage{
create_exercise_metadata(exercise_id, exercise_directory, exercise_name)
}
\arguments{
\item{exercise_id}{ID of the exercise.}

\item{exercise_directory}{Path to the directory where the exercise is downloaded into.}

\item{exercise_name}{Name of the exercise.}
}
\value{
Either NULL or an integer status. See \code{\link[base]{connections}} in the \code{base} package for details.
}
\description{
Create a JSON file containing exercie metadata.
}
\details{
Creates a JSON file in the exercise's directory root. The JSON file contains the exercise's name and id,
which is used when the user wants to upload their exercise submission to the TMC server.
}
