% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{deleteCredentials}
\alias{deleteCredentials}
\title{Delete the saved user credentials}
\usage{
deleteCredentials()
}
\value{
\code{TRUE} if succeeded in deleting the credentials.
\code{FALSE} if the deletion did not succeed or \code{.credentials}
were not found.
}
\description{
Delete the \code{.credentials} file.
}
\seealso{
\code{\link[base]{files}}
}
