% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{disable_UI_elements}
\alias{disable_UI_elements}
\title{Disable Shiny input elements using UI elements}
\usage{
disable_UI_elements(UI_list = NULL, running = FALSE)
}
\arguments{
\item{UI_list}{A named list of input element vectors or NULL.}

\item{running}{A boolean that enables Stop button when tests are
running}
}
\description{
Disable \code{Shiny} input elements given by UI elements
}
\details{
Disables the \code{Shiny} input elements.
This is done in order to prevent button actions while an
operation that might take a long itme is running.
}
\seealso{
\code{\link{enable_UI_elements}}
}
