% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{download_exercise}
\alias{download_exercise}
\title{Download an exercise from the TMC server}
\usage{
download_exercise(exercise_id,
                         zip_target = getwd(),
                         zip_name = "temp.zip",
                         exercise_directory,
                         exercise_name,
                         credentials,
                         unique_random = FALSE)
}
\arguments{
\item{exercise_id}{ID of the exercise.}

\item{zip_target}{Path to where the temporary \code{zip} file is
stored. Defaults to the current working directory.}

\item{zip_name}{Name of the temporary \code{zip} file. Default is
\code{temp.zip}.}

\item{exercise_directory}{Path to the directory where the exercise
directory is unzipped to.}

\item{exercise_name}{Name of the downloaded exercise.}

\item{credentials}{List of user credentials.}

\item{unique_random}{logical: should the name of the file be unique
and random? Default is FALSE.}
}
\value{
\code{HTTP} response to the download attempt.
}
\description{
Download an exercise from the TMC server.
}
\details{
Downloads an zipped exercise matching the given id from the TMC
server, unzips it to the given directory and creates a \code{JSON}
file containing the exercise name and id on the exercise directory.
}
\seealso{
\code{\link{getCredentials}}, \code{\link[httr]{GET}},
\code{\link[httr]{add_headers}}, \code{\link[httr]{write_disk}},
\code{\link[base]{file.remove}}, \code{\link{create_exercise_metadata}}
}
