% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_elements}
\alias{enable_elements}
\title{Enable Shiny input elements}
\usage{
enable_elements(...)
}
\arguments{
\item{...}{IDs of the \code{Shiny} input elements which are to be
enabled.}
}
\description{
Enable \code{Shiny} input elements.
}
\details{
Enables the \code{Shiny} input elements.
}
\seealso{
\code{\link[base]{as.list}},\code{\link[base]{substitute}},
\code{\link[base]{lapply}}, \code{\link[shinyjs]{enable}}
}
