% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputToggle.R
\name{enable_UI_elements}
\alias{enable_UI_elements}
\title{Enable Shiny input elements using UI list}
\usage{
enable_UI_elements(UI_list)
}
\arguments{
\item{UI_list}{A named list of vectors of input element IDs or NULL.}
}
\description{
Enable Shiny input elements using UI list
}
\details{
Enables the \code{Shiny} input elements given by UI list.

Enables the \code{Shiny} input elements.
}
\seealso{
\code{\link{disable_UI_elements}}
}
