% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{fetchClientIdAndSecret}
\alias{fetchClientIdAndSecret}
\title{Fetch client id and secret from the server}
\usage{
fetchClientIdAndSecret(serverAddress)
}
\arguments{
\item{serverAddress}{Address of the TMC server which the user wants
to log in to.}
}
\value{
Client ID and secret fetched from the server.
}
\description{
Fetch the client id and secret required in \code{OAuth2}
authentication from the server.
}
\seealso{
\code{\link[httr]{GET}}
}
