% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{getAllCourses}
\alias{getAllCourses}
\title{Get all courses offered by a TMC organization}
\usage{
getAllCourses(organization)
}
\arguments{
\item{organization}{Organization slug (Identifying URL-friendly name).}
}
\value{
List of course names and course ids belonging to the given organization.
If reading \code{.credentials.json} or sending the \code{HTTP-GET} request failed, returns
a list with 2 empty sublists called \code{id} and \code{name}.
}
\description{
Get all courses offered by a TMC organization.
}
\details{
Reads the \code{OAuth2} token and server address from \code{.credentials.json}
and uses them to make a \code{HTTP-GET} request for the list of courses belonging to
the organization.
}
\examples{
getAllCourses("hy")

}
\seealso{
\code{\link{getCredentials}}, \code{\link[httr]{stop_for_status}},
\code{\link[jsonlite]{fromJSON}}
}
