% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{getServerAddress}
\alias{getServerAddress}
\title{Get the saved server address}
\usage{
getServerAddress()
}
\value{
The saved server address from the \code{.server} file.
\code{NULL} if the file does not exist or if it is corrupted.
}
\description{
Get the saved server address from the \code{.server} file.
}
\seealso{
\code{\link[base]{file.exists}}, \code{\link[base]{scan}}
}
