% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{login}
\alias{login}
\title{Login to a TMC server}
\usage{
login(clientID, secret, username, password, serverAddress)
}
\arguments{
\item{clientID}{Client ID used in \code{OAuth2}.}

\item{secret}{Secret used in \code{OAuth2}.}

\item{username}{Username of a TMC account.}

\item{password}{Password matching the inputed username of a TMC account.}

\item{serverAddress}{Address of the TMC server which the user wants
to log in to.}
}
\value{
a named list of an \code{OAuth2} token if the authentication was succesful,
with NULL error message. Otherwise returns an error message with NULL
token.
}
\description{
Logs in to a TMC server using an username and a password
along with a clientID and secret used for \code{OAuth2}
authentication.
}
\details{
If logging in was successful, saves the login credentials
(username, access-token and the server address) in the
\code{.credentials} file.
}
\seealso{
\code{\link[httr]{POST}}, \code{\link[httr]{status_code}},
\code{\link[httr]{content}}, \code{\link{saveCredentials}}
}
