% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{processSubmission}
\alias{processSubmission}
\title{Read test result data from the submission result JSON}
\usage{
processSubmission(submitJson)
}
\arguments{
\item{submitJson}{HTTP response containing the submission result \code{JSON}.}
}
\value{
List of test results received from the submission result \code{JSON}.
}
\description{
Read test result data from the submission result \code{JSON}.
}
\details{
Creates a list of test results received from the submission
result \code{JSON}.
}
