% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercises.R
\name{sourceExercise}
\alias{sourceExercise}
\title{Source exercise .R files}
\usage{
sourceExercise(exercise)
}
\arguments{
\item{exercise}{Name of the exercise which .R files are sourced}
}
\description{
Source the .R files contained in the exercise directory's \code{R} folder.
}
