% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{submitCurrent}
\alias{submitCurrent}
\title{Submit the currently open exercise to the TMC server}
\usage{
submitCurrent()
}
\value{
Submission result \code{JSON} if processing the tests in the TMC server was
successful. \code{NULL} if processing the tests ended in error.
}
\description{
Submit the currently open exercise to the TMC server and return the
submission result \code{JSON}.
}
\details{
Reads the \code{OAuth2} token and TMC server address from
\code{.crendentials.rds} and uploads the currently open exercise to
the TMC server. If the upload was successful, starts querying the TMC server for the
submission result \code{JSON} until the server has finished processing the tests.
}
\seealso{
\code{\link{getCredentials}}, \code{\link{upload_current_exercise}},
\code{\link{getExerciseFromServer}}
}
