% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submissions.R
\name{submitCurrent}
\alias{submitCurrent}
\title{Submit the currently chosen exercise to the TMC server}
\usage{
submitCurrent(path)
}
\arguments{
\item{path}{Path to the currently chosen directory.}
}
\value{
Submission result with non \code{NULL} \code{results} if
processing the tests in the TMC server was successful. List keys:
\code{results}, \code{error}. Error is not \code{NULL} if processing
the tests ended in error.
}
\description{
Submit the currently chosen exercise to the TMC server
and return the submission result \code{JSON}.
}
\details{
Reads the \code{OAuth2} token and TMC server address from
\code{.crendentials.rds} and uploads the currently open exercise to
the TMC server. If the upload was successful, starts querying the TMC
server for the submission result \code{JSON} until the server has
finished processing the tests.
}
\seealso{
\code{\link{getCredentials}},
\code{\link{upload_current_exercise}},
\code{\link{getExerciseFromServer}}
}
