% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPQueries.R
\name{upload_exercise}
\alias{upload_exercise}
\title{Upload an exercise submission to the TMC server}
\usage{
upload_exercise(token, exercise_id, project_path, server_address,
        zip_name = "temp", remove_zip = TRUE)
}
\arguments{
\item{token}{OAuth2 token associated with the current login session.}

\item{exercise_id}{Id of the submitted exercise.}

\item{project_path}{Path to the directory of the submitted exercise.}

\item{server_address}{Address to the TMC server where the exercise is submitted to.}

\item{zip_name}{Name of the zip file which contains the exercise submission. Default is
\code{temp}}

\item{remove_zip}{\code{TRUE} or \code{FALSE} depending on if you wish to delete the
submission zip file after sending it to the server. Defaults to \code{TRUE}.}
}
\value{
HTTP response to the submission attempt.
}
\description{
Upload an exercise submission to the TMC server
}
\details{
Packs the exercise directory into a zip file and sends it to the TMC server.
}
\examples{
upload_exercise(token = token, exercise_id = 36463, project_path = getwd(),
           server_address = "https://tmc.mooc.fi")
}
